+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+------------------------
+
+## [0.3.0] - 2024-09-14
 ### Fixed
 - navigation between frames.
 - view of data type WorldTransform.
@@ -14,7 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - support patch 2.13
-- improve performance to show bytes of frames in `MEMORY`. 
+- improve performance to show bytes of frames in `MEMORY`.
 
 ### Added
 - callback in API to dynamically capture a frame using `context.Capture()`.
@@ -45,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CET api to allow other mods to track targets.
 
 <!-- Table of releases -->
-[Unreleased]: https://github.com/rayshader/cp2077-red-memorydump/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/rayshader/cp2077-red-memorydump/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/rayshader/cp2077-red-memorydump/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/rayshader/cp2077-red-memorydump/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/rayshader/cp2077-red-memorydump/releases/tag/v0.1.0